Cleaned up examples.

+++ examples/0.2/spec02individualFunctionalities/rankAndSize.cc
@@ -5,9 +5,9 @@ int darma_main(int argc, char** argv)
   darma_init(argc, argv);
 
   // get my rank
-  const int myRank = darma_spmd_rank();
+  const size_t myRank = darma_spmd_rank();
   // get size 
-  const int size = darma_spmd_size();
+  const size_t size = darma_spmd_size();
 
   std::cout << "Rank " << myRank << "/" << size << std::endl;
